+++ test/mocks/ALCXSource.sol
@@ -5,17 +5,18 @@ import {IERC20} from "openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 import {IALCXSource} from "../../src/interfaces/IALCXSource.sol";
 
 contract ALCXSource is IALCXSource {
-
     IERC20 public alcx = IERC20(0xdBdb4d16EdA451D0503b854CF79D55697F90c8DF);
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
-    function getStakeTotalDeposited(address _user, uint256) external view returns (uint256) {
+    function getStakeTotalDeposited(address _user, uint256)
+        external
+        view
+        returns (uint256)
+    {
         return balances[_user];
     }
 
-    function claim(uint256) external {
-
-    }
+    function claim(uint256) external {}
 
     function deposit(uint256, uint256 _depositAmount) external {
         alcx.transferFrom(msg.sender, address(this), _depositAmount);
